requirements.txt

mysql-connector-python
easyocr
opencv-python
numpy
pillow
matplotlib
scipy
pytesseract
Flask
Flask-SQLAlchemy
